Don't renovate any packages besides `apollo` packages on weekdays.
This would allow us to have early-warning for packages which we depend on
which start with `apollo`, which if we depend on them in our repositories,
are likely cross-repo dependencies which we want to get early warning
signals regarding their failures.

This puts off the creation of every other PR to the weekend.

Proposed for: https://github.com/apollographql/renovate-config-apollo-open-source/pull/1