Only rebase PRs on the weekends, unless they're `apollo*` packages. (#2)
* Don't renovate any packages besides `apollo` packages on weekdays.

This would allow us to have early-warning for packages which we depend on
which start with `apollo`, which if we depend on them in our repositories,
are likely cross-repo dependencies which we want to get early warning
signals regarding their failures.

This puts off the creation of every other PR to the weekend.

Proposed for: https://github.com/apollographql/renovate-config-apollo-open-source/pull/1

* Do not rebase stale PRs during the weekday, on every commit to `master`.

I think rebasing stale PRs on `master` is reasonable to do, since it is
maybe the only accurate way to put the blame directly on a specific PR
rather than the combination of that PR combined with something else.

Of course, we can use the "Retry" checkbox on the PR, but that poses
problems for PRs that update the patch or minor versions which we could
avoid being bothered with if their tests pass.  The only way to automerge
confidently seems to me to be with `rebaseStalePrs` enabled.

However, `rebaseStalePrs` is incredibly annoying when it runs during the
weekday "work hours" (a subjective period since we're global and someone is
always during work hours), which it does by default on PRs which were
already opened during the previously configured "acceptable" times.

This change, combined with the previous commit 2cdb0fff9d2ba556e42147588fac,
should make it so that ALL PRs are only renovated during their configured
times.  That means that non-`apollo` PRs are potentially rebased all the
time, but it seems we should be merging those as quickly as possible.  With
`automerge` on for those, it should happen right away, assuming the tests
pass.

However, all other packages should only be rebased according to their
schedule, which for non-`apollo` packages will be only on the weekend if
2cdb0fff9d2b is in play.

* Allow any package which has `apollo` in it, not just "starts-with"